Use int8 for upload/download priority.
GitOrigin-RevId: 3cc1c006c5ab3989be920b4d6e7e5f49aeaa42ce

+++ td/telegram/files/FileHashUploader.h
@@ -38,7 +38,7 @@ class FileHashUploader : public FileLoaderActor {
     send_closure(resource_manager_, &ResourceManager::update_resources, resource_state_);
   }
 
-  void update_priority(int32 priority) override {
+  void update_priority(int8 priority) override {
     send_closure(resource_manager_, &ResourceManager::update_priority, priority);
   }
   void update_resources(const ResourceState &other) override {
